Support using `https` requests with an `http` proxy (#28)
The error was coming from the fact that axios implementation of https requests through a http proxy is limited. In particular, it seems it doesn't correctly use the CONNECT method that allows tunneling a https request inside a http proxy.
A consequence of this is that the underlying socket connection opened to finally reach the target endpoint is not a ssl socket. This finally results in an error from the Proxy, especially with self-signed certificates.

The fix configures axios to override either http or https agent (according to whether TLS is used) with agents created by the tunnel module.

This PR was tested on a disconnected environment with an HTTP proxy, and TLS enabled with self-signed certificates.